Replace monolithic processor with state machine workflow
- Split single process into specialized workers: download, extract, upload, diarize
- Each worker handles one state transition and enqueues next step
- Added workflow orchestrator for state management and job routing
- Created systemd services for production deployment
- Updated discovery/backfill to use download queue as entry point
- Removed old worker.js and process.js
- Updated documentation for new architecture
