+++ package.json
@@ -7,6 +7,9 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "discover": "node discover.js",
     "worker": "node worker.js",
+    "worker:download": "node workers/download-worker.js",
+    "worker:extract": "node workers/extract-worker.js", 
+    "worker:upload": "node workers/upload-worker.js",
     "worker:diarize": "node diarize-worker.js",
     "backfill": "node backfill.js",
     "cleanup": "node cleanup.js",
