+++ systemd/README.md
@@ -32,11 +32,17 @@ sudo -u gnv-meetings npm install --production
 sudo systemctl daemon-reload
 
 # Enable services
-sudo systemctl enable gnv-meetings-worker.service
+sudo systemctl enable gnv-meetings-download.service
+sudo systemctl enable gnv-meetings-extract.service
+sudo systemctl enable gnv-meetings-upload.service
+sudo systemctl enable gnv-meetings-diarize.service
 sudo systemctl enable gnv-meetings-discover.timer
 
 # Start services
-sudo systemctl start gnv-meetings-worker.service
+sudo systemctl start gnv-meetings-download.service
+sudo systemctl start gnv-meetings-extract.service
+sudo systemctl start gnv-meetings-upload.service
+sudo systemctl start gnv-meetings-diarize.service
 sudo systemctl start gnv-meetings-discover.timer
 ```
 
@@ -44,20 +50,29 @@ sudo systemctl start gnv-meetings-discover.timer
 
 Check service status:
 ```bash
-sudo systemctl status gnv-meetings-worker
+sudo systemctl status gnv-meetings-download
+sudo systemctl status gnv-meetings-extract  
+sudo systemctl status gnv-meetings-upload
+sudo systemctl status gnv-meetings-diarize
 sudo systemctl status gnv-meetings-discover.timer
 ```
 
 View logs:
 ```bash
-# Worker logs
-sudo journalctl -u gnv-meetings-worker -f
+# Individual worker logs
+sudo journalctl -u gnv-meetings-download -f
+sudo journalctl -u gnv-meetings-extract -f
+sudo journalctl -u gnv-meetings-upload -f
+sudo journalctl -u gnv-meetings-diarize -f
 
 # Discovery logs
 sudo journalctl -u gnv-meetings-discover -f
 
+# All workers combined
+sudo journalctl -u gnv-meetings-download -u gnv-meetings-extract -u gnv-meetings-upload -u gnv-meetings-diarize -f
+
 # Last 100 lines
-sudo journalctl -u gnv-meetings-worker -n 100
+sudo journalctl -u gnv-meetings-download -n 100
 ```
 
 Check timer schedule:
